requirements.txt
requests
beautifulsoup4
python-dotenv
supabase
openai
pydantic
tldextract
readability-lxml
lxml
